+++ semicolon/src/main/java/dukku/semicolon/shared/payment/dto/PaymentConfirmResponse.java
@@ -108,16 +108,16 @@ public class PaymentConfirmResponse {
         /**
          * 최종 결제 금액
          */
-        private Integer finalPayAmount;
+        private Long finalPayAmount;
 
         /**
          * 예치금 사용 금액
          */
-        private Integer depositUseAmount;
+        private Long depositUseAmount;
 
         /**
          * PG 결제 금액
          */
-        private Integer pgPayAmount;
+        private Long pgPayAmount;
     }
 }
